Spinboxes in kmeans dialog are updated when points are imported
Spinboxes for min, max, numOfPoints and dimension are updated
once the points are imported from a file.

+++ KMeansDialog.h
@@ -26,7 +26,7 @@ public slots:
   void getPrevStep();
   void importPoints();
   void updateImportedPointParameters(int numOfPoint, int dimension,
-                                     int min, int max);
+                                     float min, float max);
   void updatePointInfoLabel(QString s);
 
 signals:
